Refactor algorithm spi
