+++ features/mask/api/src/main/java/org/apache/shardingsphere/mask/spi/MaskAlgorithm.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.mask.spi;
 
-import org.apache.shardingsphere.infra.util.spi.type.typed.TypedSPI;
+import org.apache.shardingsphere.infra.util.spi.type.typed.algorithm.ShardingSphereAlgorithm;
 
 /**
  * Mask algorithm.
@@ -25,7 +25,7 @@ import org.apache.shardingsphere.infra.util.spi.type.typed.TypedSPI;
  * @param <I> type of plain value
  * @param <O> type of masked value
  */
-public interface MaskAlgorithm<I, O> extends TypedSPI {
+public interface MaskAlgorithm<I, O> extends ShardingSphereAlgorithm {
     
     /**
      * Mask.
